requirements.txt
PyPDF2-3.0.1
pdfplumber==0.10.2
